requirements.txt
# Django e componentes principais
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
django-filter==23.5
django-allauth==0.61.0
django-oauth-toolkit==2.3.0
django-environ==0.11.2
django-storages==1.14.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-redis==5.4.0

# Banco de dados
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Autenticação e segurança
PyJWT==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Processamento assíncrono
celery==5.3.6
redis==5.0.1
amqp==5.2.0
kombu==5.3.4
flower==2.0.1

# Web scraping e HTTP
httpx==0.26.0
selectolax==0.3.17
beautifulsoup4==4.12.3
lxml==5.1.0
requests==2.31.0

# Processamento de documentos e OCR
pytesseract==0.3.10
Pillow==10.2.0
opencv-python==4.9.0.80
pdf2image==1.17.0
PyMuPDF==1.23.21
python-docx==1.1.0
reportlab==4.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
docxtpl==0.16.7

# NLP e processamento de texto
spacy==3.7.2
nltk==3.8.1

# Integração com Google Drive
google-api-python-client==2.114.0
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1

# Blockchain (para assinatura digital)
web3==6.11.3
py-solc-x==1.1.1

# Servidor web para produção
gunicorn==21.2.0
uvicorn==0.27.1
whitenoise==6.6.0

# Ferramentas de desenvolvimento
python-dotenv==1.0.0
pytest==7.4.3
pytest-django==4.7.0
coverage==7.4.1
black==23.12.1
flake8==6.1.0
isort==5.13.2
pre-commit==3.6.0
django-debug-toolbar==4.2.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3.post1
Markdown==3.5.2
tqdm==4.66.1
pydantic==2.5.3
